Added per request-response deadline with lodash.
This follows the instruction of the dev in the github issue #9443.
The call.cancel() throws an uncaught exception. I can only handle it as an uncaughtException as far as I know.
I'll ask the dev about that too.

The server now uses a basic normal distribution of simulated work delays.